feat(maprender): Phase 7a TrackingStationScene + TS shadow parity
Extracts the scene-agnostic resolve/body/orbit plumbing into GhostMapSceneBase
(MapViewScene collapses to just the FLIGHT scene gate) and adds
TrackingStationScene (TRACKSTATION gate). ParsekTrackingStation.Update now runs
ShadowRenderDriver.RunFrame over the TS map ghosts (gated on mapRenderTracing,
try/catch), so the Phase-6 reconciler reports decision-vs-old-truth parity in
the Tracking Station as well as flight - writing nothing to stock either way.

TS renders a single span instance (the span clock already enforces that, no
overlap machinery). 7b (make-before-break / cold-start / per-scene patched-conic
divergence, gated on the §15.2 in-game probe) remains. Full suite green (13269).
